fix: decouple single gas price source response
