+++ crates/services/gas_price_service/src/fuel_gas_price_updater/fuel_da_source_adapter/dummy_source.rs
@@ -1,14 +1,14 @@
 use crate::fuel_gas_price_updater::{
     service::Result as DaGasPriceSourceResult,
+    DaGasPrice,
     DaGasPriceSource,
-    RawDaGasPrice,
 };
 
 pub struct DummyDaGasPriceSource;
 
 #[async_trait::async_trait]
 impl DaGasPriceSource for DummyDaGasPriceSource {
-    async fn get(&mut self) -> DaGasPriceSourceResult<RawDaGasPrice> {
-        Ok(RawDaGasPrice::default())
+    async fn get(&mut self) -> DaGasPriceSourceResult<DaGasPrice> {
+        Ok(DaGasPrice::default())
     }
 }
